DEMRUM-774 Updates per review comments
